feat: generate all needed mappings
